# Box3: optimized matrix transform - ApplyMatrix4