Little known optimization for AABB matrix transformation.

Instead of performing a Vec3/Matrix multiplication 8 times (one for each corner), we instead perform only 2. 
Once for the center of the box, and a direction transformation (normal-like) for the extent (half-size). 
Then we reconstruct the box bounds back in the correct coordinate space.

[AABB/Matrix Optimization Blog](https://zeux.io/2010/10/17/aabb-from-obb-with-component-wise-abs/) by Arseny Kapoulkine (Zeux)

Tests, on different devices, show a 3-5x performance gain.

| | 10k  | 100k | 500k | 1M |
| --- | --------| ------ | --------| --- |
| original | 1.33 ms  | 8.34 ms  | 39.58 ms  | 79.08 ms |
| new | 0.37 ms  | 2.13 ms  | 9.15 ms | 17.33 ms |
| ratio | 3.59 | 3.91 | 4.32 | 4.56 |