@sciecode 2024-05-12
Unfortunately this won't work. It only supports affine-transformations in its current form.

A similar attempt to change the `Box3.ApplyMatrix4` was already made & reverted, due to #14655.
So I imagine the same thing would happen here.

Perhaps we could introduce an affine only transformation for this? 

I don't seen many reasons to transform a `Box3` to NDC. Since it's very easy to end up with `NaN`s and `Infinity` due to perspective divide. Even for the cases where users want a simpler frustum culling, clip-space is usually used (currently impossible), not NDC. Eitherway, using an affine-transform is likely much more common, and having a specialized function that doesn't incur a 4x performance penalty is probably desirable.

@WestLangley @Mugen87 thoughts?