@WestLangley 2024-05-12
>Perhaps we could introduce an affine only transformation for this?

I think it would be better to test if the matrix is affine, and use the proposed algorithm if it is, and the existing algorithm if it is not.

But I would only do that if there is a compelling use case to justify doing so.