@sciecode 2024-05-13
I can provide a few recurring uses for my projects. 

I often rewrite a custom frustum culling system, which transforms object's AABB to world-space, independant of the default system. I do this for one of two reasons: 

Either because the default `sphere` bounding volume isn't well suited for tiled/grids/octree based entities, and a bounding-box provides tighter culling for potentially expensive render entities. Or because the DAG-based scene traversal is extremelly slow for a large number of nested entities, and it can be more efficiently computed by flattening the traversal and/or use Object AABB's that don't recurse all the way down.

Another common usage has been when building alternative raycasting systems that also don't suffer from the default system traversal / excess of unnecessary information.

I did a quick search and every single instance of `Box3.ApplyMatrix4` internal to three.js (although there aren't many), do use it with affine-transformation.

---- 

Also worth mentioning, I already use the proposed method for my own uses. In fact the only reason I noticed that `ApplyMatrix4` wasn't implemented this way, was because I mistakenly used it in my own project and noticed a 5x performance degradation of a very hot loop. So I decided to share this with others that might also be running into this.