@Mugen87 2024-05-21
I would favor a single code path and not multiple ones depending on the input matrix. I have the feeling this would overcomplicate things.