@sciecode 2024-05-21
I agree. I would prefer to introduce a separate function `applyAffine` which makes use of the propose method. And we update methods that are guaranteed to only call this method with affine-transforms, so that they may make use of the fast path (like `setFromObject`).

If that's acceptable, I'll patch the PR. Otherwise, let me know so I can close it.